examples/gpu-examples: move LoadShader() to its own, internal package
